// src/utils/types/instrument.tsx
export type FormValues = {
    description: string | undefined;
    code: string | undefined;
    calibrationFrequency: number | undefined;
    manufacturer: string | undefined;
    additionalReference1: string | undefined;
    additionalReference2: string | undefined;
    additionalReference3: string | undefined;
    acquisitionDate: string | undefined;
    inventory: string | undefined;
    supplier: string | undefined;
    serieNumber: string | undefined;
    familyID: string | undefined;
    family: string | undefined;
    supplierDescription: string | undefined;
    situation: string | undefined;
    status: string | undefined;
    situationReason: string | undefined;
    situationJustification: string | undefined;
  };